feat: add repository card skeleton
