+++ front-end/src/pages/public/LandingPage/SearchSection.tsx
@@ -1,18 +1,19 @@
-import { Search, Loader2 } from "lucide-react";
+import { Search } from "lucide-react";
 import { Input } from "../../../components/ui/input";
 import { useState } from "react";
 import { Button } from "../../../components/ui/button";
 import { useDebounce } from "@/lib/useDebounce";
 import { useQuery } from "@tanstack/react-query";
-import { RepositoryCard } from "../../../components/RepositoryCard";
+import { RepositoryCard } from "../../../components/RespositoryCard";
 import { RepositoriesData, Repository } from "@/types/repository";
 import { fetchRepositories } from "@/lib/fetchRepositories";
+import { Skeleton } from "@/components/RespositoryCard/skeleton";
 
 export default function SearchSection() {
   const [searchInput, setSearchInput] = useState<string>("");
   const debouncedSearch = useDebounce(searchInput, 500);
   const [currentPage, setCurrentPage] = useState<number>(1);
-  const perPage = 3;
+  const perPage = 6;
 
   const { data, isLoading, error } = useQuery<RepositoriesData | undefined>({
     queryKey: ["repositories", debouncedSearch, currentPage],
@@ -26,9 +27,9 @@ export default function SearchSection() {
   return (
     <section className="flex flex-col">
       <div className="space-y-4 mb-6">
-        <h2 className="text-muted-foreground max-w-2xl">
-          Descubra projetos open source que combinam com seus interesses e nível
-          de habilidade
+        <h2 className="text-muted-foreground max-w-2xl font-medium">
+          Discover open source projects that match your interests and skill
+          level
         </h2>
       </div>
 
@@ -37,7 +38,7 @@ export default function SearchSection() {
           <Search className="absolute left-3 top-1/2 -translate-y-1/2 text-muted-foreground w-4 h-4" />
           <Input
             className="pl-10"
-            placeholder="Buscar repositórios..."
+            placeholder="Search repositories..."
             value={searchInput}
             onChange={(e) => setSearchInput(e.target.value)}
           />
@@ -45,8 +46,10 @@ export default function SearchSection() {
       </div>
 
       {isLoading ? (
-        <div className="flex justify-center items-center py-12">
-          <Loader2 className="h-8 w-8 animate-spin text-primary" />
+        <div className="grid gap-8 md:grid-cols-3 pb-6">
+          {new Array(6).fill(null).map(() => (
+            <Skeleton />
+          ))}
         </div>
       ) : error ? (
         <div className="text-center py-8 text-destructive">
@@ -56,7 +59,7 @@ export default function SearchSection() {
         <div className="space-y-4 mb-6">
           <div className="grid gap-8 md:grid-cols-3">
             {repositories.map((repository: Repository) => (
-              <RepositoryCard repo={repository} />
+              <RepositoryCard repository={repository} />
             ))}
           </div>
         </div>
